+++ src/main/java/org/example/fashion_api/Repositories/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<Account, Integer> {
+public interface UserRepo extends JpaRepository<Account, Long> {
     Optional<Account> findByUsername(String username);
 }
